+++ src/xyz/oihana/schema/constants/traits/StatisticsTrait.php
@@ -5,6 +5,8 @@ namespace xyz\oihana\schema\constants\traits;
 use xyz\oihana\schema\constants\traits\statistics\CustomerStatisticsTrait;
 use xyz\oihana\schema\constants\traits\statistics\HasTradingMeasuresTrait;
 use xyz\oihana\schema\constants\traits\statistics\ObservationSeriesTrait;
+use xyz\oihana\schema\constants\traits\statistics\SalesObjectivesTrait;
+use xyz\oihana\schema\constants\traits\statistics\SellerStatisticsTrait;
 use xyz\oihana\schema\constants\traits\statistics\StatisticsRecordTrait;
 
 /**
@@ -19,5 +21,7 @@ trait StatisticsTrait
     use CustomerStatisticsTrait ,
         HasTradingMeasuresTrait ,
         ObservationSeriesTrait  ,
+        SalesObjectivesTrait    ,
+        SellerStatisticsTrait   ,
         StatisticsRecordTrait   ;
 }