+++ src/xyz/oihana/schema/constants/traits/statistics/SalesObjectivesTrait.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\statistics;
+
+/**
+ * The property name constants of the {@see \xyz\oihana\schema\statistics\SalesObjectives} class.
+ *
+ * @package xyz\oihana\schema\constants\traits\statistics
+ * @author  Marc Alcaraz (eKameleon)
+ * @since   1.5.0
+ */
+trait SalesObjectivesTrait
+{
+    const string ASSIGNED_CATEGORY = 'assignedCategory' ;
+    const string ASSIGNED_CUSTOMER = 'assignedCustomer' ;
+}