+++ src/xyz/oihana/schema/constants/traits/statistics/SellerStatisticsTrait.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\statistics;
+
+/**
+ * The property name constants of the {@see \xyz\oihana\schema\statistics\SellerStatistics} class.
+ *
+ * @package xyz\oihana\schema\constants\traits\statistics
+ * @author  Marc Alcaraz (eKameleon)
+ * @since   1.5.0
+ */
+trait SellerStatisticsTrait
+{
+    const string ASSIGNED_CUSTOMER = 'assignedCustomer' ;
+}